requirements.txt
sqlalchemy
wxPython
pypubsub